@@ -5,7 +5,7 @@
 
 package agent
 
-var DefaultBanner []byte = []byte{
+var DefaultBanner = []byte{
 	0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a, 0x00, 0x00, 0x00, 0x0d,
 	0x49, 0x48, 0x44, 0x52, 0x00, 0x00, 0x01, 0x70, 0x00, 0x00, 0x01, 0xae,
 	0x08, 0x06, 0x00, 0x00, 0x00, 0x89, 0x2d, 0x5d, 0x60, 0x00, 0x00, 0x00,
