@@ -62,10 +62,10 @@ func (b *Bus) Initialize() {
 				err := fmt.Sprintf("Provider %s at %s had an error: %s", p.Name, p.Executable, r.Error)
 				fmt.Println(err)
 				os.Exit(1)
-			} else {
-				if r.State != "" {
-					fmt.Println(fmt.Sprintf("[provider event: %s]", e), r.State)
-				}
+			}
+
+			if r.State != "" {
+				fmt.Println(fmt.Sprintf("[provider event: %s]", e), r.State)
 			}
 		})
 	}
